using blueprint to split out routes like user, admin, and general
