+++ drinkorder/admins/__init__.py
@@ -1,0 +1,2 @@
+# In order to tell Python that this folder is a package and it doesn't even have to contain anything, 
+# it can be an empty file and Python will still know that is the package
